Squashed commit of the NoHybridConventions branch
commit 8ab033088358908a0d0bad2610838d9a6104f094
Author: Deividas <deividas@deividas-HP-ProBook-450-G1.(none)>
Date:   Mon Nov 17 18:43:16 2014 +0200

    i like this name better for constant

    ...

commit 074bd11541b7ce4aa8512aa5f97021de19f8dc8c
Author: Deividas <deividas@deividas-HP-ProBook-450-G1.(none)>
Date:   Mon Nov 17 17:02:47 2014 +0200

    one by one

+++ util/spec/DefaultHashes.h
@@ -89,7 +89,7 @@ struct Hasher<std::string> {
     size_t operator()(const std::string& t) {
         // DUMMY, BAD FUNCTION
         size_t res = 0;
-        for (int i = 0; i < t.size(); ++i) {
+        for (size_t i = 0; i < t.size(); ++i) {
             res += t[i];
         }
         return res;
